refactor: consolidate anime/manga category screens into shared composable (R17)
- Created shared CategoryScreen composable extracting ~100 lines of duplicated UI logic
- AnimeCategoryScreen and MangaCategoryScreen now thin wrappers (26 lines each)
- Zero behavioral changes - exact same UI and interactions
- Net: -88 lines (3 files: 1 created, 2 simplified)
