+++ app/src/main/java/eu/kanade/presentation/category/AnimeCategoryScreen.kt
@@ -1,30 +1,13 @@
 package eu.kanade.presentation.category
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyListState
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.toMutableStateList
-import androidx.compose.ui.Modifier
-import eu.kanade.presentation.category.components.CategoryFloatingActionButton
-import eu.kanade.presentation.category.components.CategoryListItem
 import eu.kanade.tachiyomi.ui.category.anime.AnimeCategoryScreenState
-import sh.calvin.reorderable.ReorderableItem
-import sh.calvin.reorderable.rememberReorderableLazyListState
 import tachiyomi.domain.category.model.Category
-import tachiyomi.i18n.MR
-import tachiyomi.presentation.core.components.material.Scaffold
-import tachiyomi.presentation.core.components.material.padding
-import tachiyomi.presentation.core.screens.EmptyScreen
 
+/**
+ * Anime-specific wrapper for the shared CategoryScreen composable.
+ * Delegates to the shared implementation while maintaining the original API.
+ */
 @Composable
 fun AnimeCategoryScreen(
     state: AnimeCategoryScreenState.Success,
@@ -34,80 +17,13 @@ fun AnimeCategoryScreen(
     onClickDelete: (Category) -> Unit,
     onChangeOrder: (Category, Int) -> Unit,
 ) {
-    val lazyListState = rememberLazyListState()
-    Scaffold(
-        floatingActionButton = {
-            CategoryFloatingActionButton(
-                lazyListState = lazyListState,
-                onCreate = onClickCreate,
-            )
-        },
-    ) { paddingValues ->
-        if (state.isEmpty) {
-            EmptyScreen(
-                stringRes = MR.strings.information_empty_category,
-                modifier = Modifier.padding(paddingValues),
-            )
-            return@Scaffold
-        }
-
-        CategoryContent(
-            categories = state.categories,
-            lazyListState = lazyListState,
-            paddingValues = paddingValues,
-            onClickRename = onClickRename,
-            onClickHide = onClickHide,
-            onClickDelete = onClickDelete,
-            onChangeOrder = onChangeOrder,
-        )
-    }
+    CategoryScreen(
+        categories = state.categories,
+        isEmpty = state.isEmpty,
+        onClickCreate = onClickCreate,
+        onClickRename = onClickRename,
+        onClickHide = onClickHide,
+        onClickDelete = onClickDelete,
+        onChangeOrder = onChangeOrder,
+    )
 }
-
-@Composable
-private fun CategoryContent(
-    categories: List<Category>,
-    lazyListState: LazyListState,
-    paddingValues: PaddingValues,
-    onClickRename: (Category) -> Unit,
-    onClickHide: (Category) -> Unit,
-    onClickDelete: (Category) -> Unit,
-    onChangeOrder: (Category, Int) -> Unit,
-) {
-    val categoriesState = remember { categories.toMutableStateList() }
-    val reorderableState = rememberReorderableLazyListState(lazyListState, paddingValues) { from, to ->
-        val item = categoriesState.removeAt(from.index)
-        categoriesState.add(to.index, item)
-        onChangeOrder(item, to.index)
-    }
-
-    LaunchedEffect(categories) {
-        if (!reorderableState.isAnyItemDragging) {
-            categoriesState.clear()
-            categoriesState.addAll(categories)
-        }
-    }
-
-    LazyColumn(
-        modifier = Modifier.fillMaxSize(),
-        state = lazyListState,
-        contentPadding = PaddingValues(MaterialTheme.padding.medium),
-        verticalArrangement = Arrangement.spacedBy(MaterialTheme.padding.small),
-    ) {
-        items(
-            items = categoriesState,
-            key = { category -> category.key },
-        ) { category ->
-            ReorderableItem(reorderableState, category.key) {
-                CategoryListItem(
-                    modifier = Modifier.animateItem(),
-                    category = category,
-                    onRename = { onClickRename(category) },
-                    onHide = { onClickHide(category) },
-                    onDelete = { onClickDelete(category) },
-                )
-            }
-        }
-    }
-}
-
-private val Category.key inline get() = "category-$id"
